feat: add test addPromocode function
